increase TTKey size

+++ src/types.h
@@ -7,7 +7,7 @@
 // define bitboard data type
 using Bitboard = uint64_t;
 // define poskey data type
-using TTKey = uint16_t;
+using TTKey = uint32_t;
 // define poskey data type
 using ZobristKey = uint64_t;
